requirements.txt
python-telegram-bot>=20.0,<21.0
tweepy>=4.14.0,<5.0
groq>=0.4.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
requests>=2.28.0
loguru>=0.6.0
schedule>=1.1.0
tenacity>=8.0.0
pyrogram>=2.0.0